+++ src/templates/header.html
@@ -14,10 +14,6 @@
             <a href="wishList.php" id="wishList"><span>Wish list</span></a>
             <a href="login.php" id="loginForm"><span>Login</span></a>
     </div>
-<input type="checkbox" id="check">
-<label for="check" class="checkbtn">
-    <i class="fas fa-bars">&#9776</i>
-</label>
 <nav id="menu">
     <ul>
         <li lang="en">Home</li>
